backend/requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
mysqlclient
werkzeug
